Add files via upload
The Arduino program file has the script in it, but I also included the script individually if you just want to take a look at the code. That being said, the Arduino program file has the old runs with the older version of the Arduino program, but it gives you an idea of how the CSV should look.