Add files via upload
Okay, this is the one where I tried to figure out the \x00 problem with converting to floats. This just did not work, so next attempt I am going to just write to one big file and split it up later. REMEMBER, the ython script is ONLY FOR DATA ANALYSIS. theres no point in doing the analysis fast, only the data writing.