Add files via upload
This is the code with multiprocessing and classes only for finding the analysis section. The classes include: find zeros, which just find the zeros and put them in their own file called cross times, and the make plot, which is self-explanatory. Other than that the multiprocessing addition will make it take about .7 seconds.