Add files via upload
Okay, this just takes out the multiprocessing because for some reason it was causing it to run the entire program MULTIPLE times. It did have an error that caused it to repeat the cross items twice but that was because it had the write row part in the for a loop. Other than that I created a functions file that will shorten the actual code (and is just cool). I'll go back to multiprocessing but with the fix in the find zeros folder it'll work I just don't want to run it while it's still so buggy. Also, the entire code is in a constant try-catch and will not break if there is an error now. Nice!